requirements.txt
asgiref==3.5.2
Django==4.0.5
djangorestframework==3.13.1
mysql-client==0.0.1
mysql-connector-python==8.0.29
mysqlclient==2.1.1
Pillow==9.2.0
protobuf==4.21.2
psycopg2==2.9.3
psycopg2-binary==2.9.2
pytz==2021.1
sqlparse==0.4.2
